# groucho-lsh-rest/api.yaml
swagger: "2.0"
info:
  description: "Groucho REST API for Monitoring Object State "
  version: "1.0.API.SNAPSHOT"
  title: "Groucho State Observer"
  termsOfService: "http://saks-wiki.iasi.cnr.it/groucho"
  contact:
    email: "guglielmo.deangelis@iasi.cnr.it"
  license:
    name: "GNU General Public License (GPL ver.3)"
    url: "http://www.gnu.org/licenses/gpl.txt"
host: "localhost"
port: "8080"
basePath: "/groucho-state-observer"
tags:
- name: "isStateUnknown"
  description: "Check if a given state has been previously observed"
  externalDocs:
    description: "Find out more"
    url: "http://saks-wiki.iasi.cnr.it/groucho"
- name: "markState"
  description: "Assert that a given state has been observed"
  externalDocs:
    description: "Find out more"
    url: "http://saks-wiki.iasi.cnr.it/groucho"
- name: "resetStateObserver"
  description: "Suggest to Restore the State of the Observer to the Default Configurations. It is only a suggestion, there is no commitment from the receiver"
  externalDocs:
    description: "Find out more"
    url: "http://saks-wiki.iasi.cnr.it/groucho"
schemes:
- "http"
paths:
  /isunknown:
    get:
      tags:
      - "isStateUnknown"
      summary: "Check if a given state has been previously observed"
      description: "Returns TRUE if a State has been never observed, FALSE otherwise"
      operationId: "isStateUnknown"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: "body"
          in: "body"
          description: "The Actual Internal Representation as a String of the Considered State"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: "Successful Operation."
          schema:
            format: "boolean"
        400:
          description: "Invalid State Supplied."
        404:
          description: "Check Exception, State not found."
  /isunknown/{stateStringLSH}:
    get:
      tags:
      - "isStateUnknown"
      summary: "Check if a given state has been previously observed"
      description: "Returns TRUE if a State has been never observed, FALSE otherwise. The input parameter is supposed to be a codification in LSH of the actual internal state representation."
      operationId: "isStateUnknownLSH"
      produces:
      - "application/json"
      parameters:
      - name: "stateStringLSH"
        in: "path"
        description: "LSH String of Considered State."
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            format: "boolean"
        400:
          description: "Invalid LSH State Supplied"
        404:
          description: "Check Exception, LSH State not found"
  /reset:
    put:
      tags:
      - "resetStateObserver"
      summary: "Suggest to Restore the State of the Observer to the Default Configurations."
      description: "Suggest to Restore the State of the Observer to the Default Configurations. It is only a suggestion, there is no commitment from the reciver."
      operationId: "resetStateObserver"
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful Operation."
          schema:
            format: "boolean"
        400:
          description: "Reset Not Allowed."
        404:
          description: "Reset Exception."
  /mark:
    post:
      tags:
      - "markState"
      summary: "Assert that a given state has been observed."
      description: "Assert that a given state has been observed. It return TRUE is the state has been marked, FALSE otherwise."
      operationId: "markState"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "The Actual Internal Representation as a String of the Considered State."
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Successful Operation."
          schema:
            format: "boolean"
        400:
          description: "Invalid State Supplied."
        404:
          description: "Mark Exception."
  /mark/{stateStringLSH}:
    put:
      tags:
      - "markState"
      summary: "Assert that a given state has been observed."
      description: "Assert that a given state has been observed. It return TRUE is the state has been marked, FALSE otherwise. The input parameter is supposed to be a codification in LSH of the actual internal state representation."
      operationId: "markStateLSH"
      produces:
      - "application/json"
      parameters:
      - name: "stateStringLSH"
        in: "path"
        description: "LSH String of Considered State"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation."
          schema:
            format: "boolean"
        400:
          description: "Invalid LSH State Supplied."
        404:
          description: "Mark Exception."
externalDocs:
  description: "Find out more"
  url: "http://saks-wiki.iasi.cnr.it/groucho"